Restructured Repo structure for updated code.
FeederMonitor is the refactored version of Dr. Heidorn's original code
with the included implementation of hardware interrupts for Power Saving
in the feeders.

The original code has been moved in to the pre_interrupt_code directory.

The "Example Data" directory will eventually contain example data but
currently contains test files.